apresentacao/requirements.txt
# Análise CISARP - Dependências Python
# Instalar com: pip install -r requirements.txt

# Manipulação de dados
pandas==2.1.0
numpy==1.24.3
openpyxl==3.1.2

# Banco de dados
psycopg2-binary==2.9.7

# Visualização
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Estatística
scipy==1.11.2

# Relatórios
reportlab==4.0.4
